requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Document processing
PyMuPDF==1.24.13
python-docx==1.1.2
pdfplumber==0.11.4
Pillow==11.0.0

# OCR
pytesseract==0.3.13
easyocr==1.7.2

# AI / ML
anthropic==0.40.0

# File type detection
python-magic==0.4.27

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.17
aiofiles==24.1.0

# Security
cryptography==43.0.3
